+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lastshotlabs/slingshot
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @lastshotlabs/slingshot-community@0.2.6
+
 ## 0.2.6
 
 ### Patch Changes